move mina and utils stuff into their folders
