+++ src/lib/circuit.ts
@@ -4,7 +4,10 @@ import { MlFieldArray, MlFieldConstArray } from './ml/fields.js';
 import { withThreadPool } from '../snarky.js';
 import { Provable } from './provable.js';
 import { snarkContext, gatesFromJson } from './provable-context.js';
-import { prettifyStacktrace, prettifyStacktracePromise } from './errors.js';
+import {
+  prettifyStacktrace,
+  prettifyStacktracePromise,
+} from './util/errors.js';
 import { ProvablePure } from './provable-types/provable-intf.js';
 
 // external API
