+++ src/lib/field.ts
@@ -5,7 +5,7 @@ import { defineBinable } from '../bindings/lib/binable.js';
 import type { NonNegativeInteger } from '../bindings/crypto/non-negative.js';
 import { asProver, inCheckedComputation } from './provable-context.js';
 import { Bool } from './bool.js';
-import { assert } from './errors.js';
+import { assert } from './util/errors.js';
 import { Provable } from './provable.js';
 import {
   assertEqual,
