+++ src/lib/mina/base58-encodings.ts
@@ -1,5 +1,5 @@
-import { fieldEncodings } from './base58.js';
-import { Field } from './core.js';
+import { fieldEncodings } from '../util/base58.js';
+import { Field } from '../core.js';
 
 export { TokenId, ReceiptChainHash, LedgerHash, EpochSeed, StateHash };
