+++ src/lib/mina/errors.ts
@@ -1,5 +1,5 @@
 import { Types } from '../../bindings/mina-transaction/types.js';
-import { TokenId } from '../account-update.js';
+import { TokenId } from './account-update.js';
 import { Int64 } from '../int.js';
 
 export { invalidTransactionError };
