+++ src/lib/mina/events.ts
@@ -1,9 +1,9 @@
-import { prefixes } from '../bindings/crypto/constants.js';
-import { prefixToField } from '../bindings/lib/binable.js';
+import { prefixes } from '../../bindings/crypto/constants.js';
+import { prefixToField } from '../../bindings/lib/binable.js';
 import {
   GenericProvableExtended,
   GenericSignableField,
-} from '../bindings/lib/generic.js';
+} from '../../bindings/lib/generic.js';
 
 export { createEvents, dataAsHash };
