+++ src/lib/mina/graphql.ts
@@ -1,6 +1,6 @@
 import { UInt32 } from '../int.js';
-import type { ZkappCommand } from '../account-update.js';
-import type { ActionStatesStringified } from '../fetch.js';
+import type { ZkappCommand } from './account-update.js';
+import type { ActionStatesStringified } from './fetch.js';
 import { Types } from '../../bindings/mina-transaction/types.js';
 
 export {
