+++ src/lib/mina/mina-instance.ts
@@ -6,10 +6,10 @@ import { UInt64, UInt32 } from '../int.js';
 import { PublicKey, PrivateKey } from '../signature.js';
 import type { EventActionFilterOptions } from '././../mina/graphql.js';
 import type { NetworkId } from '../../mina-signer/src/types.js';
-import type { Transaction, PendingTransaction } from '../mina.js';
+import type { Transaction, PendingTransaction } from './mina.js';
 import type { Account } from './account.js';
-import type { NetworkValue } from '../precondition.js';
-import type * as Fetch from '../fetch.js';
+import type { NetworkValue } from './precondition.js';
+import type * as Fetch from './fetch.js';
 
 export {
   Mina,
