+++ src/lib/mina/smart-contract-base.ts
@@ -2,7 +2,7 @@
  * This file exists to avoid an import cycle between code that just needs access
  * to a smart contract base class, and the code that implements `SmartContract`.
  */
-import type { SmartContract } from '../zkapp.js';
+import type { SmartContract } from './zkapp.js';
 
 export { isSmartContract, SmartContractBase };
