+++ src/lib/mina/smart-contract-context.ts
@@ -1,6 +1,6 @@
-import type { SmartContract } from '../zkapp.js';
-import type { AccountUpdate, AccountUpdateLayout } from '../account-update.js';
-import { Context } from '../global-context.js';
+import type { SmartContract } from './zkapp.js';
+import type { AccountUpdate, AccountUpdateLayout } from './account-update.js';
+import { Context } from '../util/global-context.js';
 import { currentTransaction } from './transaction-context.js';
 
 export { smartContractContext, SmartContractContext, accountUpdateLayout };
