+++ src/lib/mina/state.ts
@@ -1,13 +1,13 @@
-import { FlexibleProvablePure } from './provable-types/struct.js';
+import { FlexibleProvablePure } from '../provable-types/struct.js';
 import { AccountUpdate, TokenId } from './account-update.js';
-import { PublicKey } from './signature.js';
+import { PublicKey } from '../signature.js';
 import * as Mina from './mina.js';
 import { fetchAccount, networkConfig } from './fetch.js';
 import { SmartContract } from './zkapp.js';
-import { Account } from './mina/account.js';
-import { Provable } from './provable.js';
-import { Field } from '../lib/core.js';
-import { ProvablePure } from './provable-types/provable-intf.js';
+import { Account } from './account.js';
+import { Provable } from '../provable.js';
+import { Field } from '../core.js';
+import { ProvablePure } from '../provable-types/provable-intf.js';
 
 // external API
 export { State, state, declareState };
