+++ src/lib/mina/token/forest-iterator.ts
@@ -3,7 +3,7 @@ import {
   AccountUpdateForest,
   AccountUpdateTreeBase,
   TokenId,
-} from '../../account-update.js';
+} from '../account-update.js';
 import { Field } from '../../core.js';
 import { Provable } from '../../provable.js';
 import { Struct } from '../../provable-types/struct.js';
