+++ src/lib/mina/token/forest-iterator.unit-test.ts
@@ -6,7 +6,7 @@ import {
   AccountUpdateForest,
   TokenId,
   hashAccountUpdate,
-} from '../../account-update.js';
+} from '../account-update.js';
 import { TypesBigint } from '../../../bindings/mina-transaction/types.js';
 import { Pickles } from '../../../snarky.js';
 import {
