+++ src/lib/mina/token/token-methods.ts
@@ -1,7 +1,7 @@
-import { AccountUpdate, Authorization, TokenId } from '../../account-update.js';
+import { AccountUpdate, Authorization, TokenId } from '../account-update.js';
 import { isSmartContract } from '../smart-contract-base.js';
 import { PublicKey } from '../../signature.js';
-import type { SmartContract } from '../../zkapp.js';
+import type { SmartContract } from '../zkapp.js';
 import { UInt64 } from '../../int.js';
 import { Bool, Field } from '../../core.js';
