+++ src/lib/mina/transaction.ts
@@ -7,17 +7,17 @@ import {
   addMissingSignatures,
   TokenId,
   addMissingProofs,
-} from '../account-update.js';
+} from './account-update.js';
 import { Field } from '../core.js';
 import { PrivateKey, PublicKey } from '../signature.js';
 import { UInt32, UInt64 } from '../int.js';
 import { Empty, Proof } from '../proof-system.js';
 import { currentTransaction } from './transaction-context.js';
 import { Provable } from '../provable.js';
-import { assertPreconditionInvariants } from '../precondition.js';
+import { assertPreconditionInvariants } from './precondition.js';
 import { Account } from './account.js';
 import { type FeePayerSpec, activeInstance } from './mina-instance.js';
-import * as Fetch from '../fetch.js';
+import * as Fetch from './fetch.js';
 import { type SendZkAppResponse, sendZkappQuery } from './graphql.js';
 import { type FetchMode } from './transaction-context.js';
 import { assertPromise } from '../util/assert.js';
