+++ src/lib/ml/consistency.unit-test.ts
@@ -2,7 +2,7 @@ import { Test } from '../../snarky.js';
 import { Random, test } from '../testing/property.js';
 import { Field, Bool } from '../core.js';
 import { PrivateKey, PublicKey } from '../signature.js';
-import { TokenId, dummySignature } from '../account-update.js';
+import { TokenId, dummySignature } from '../mina/account-update.js';
 import { Ml } from './conversion.js';
 import { expect } from 'expect';
 import { FieldConst } from '../provable-core/fieldvar.js';
