+++ src/lib/provable-context.ts
@@ -1,7 +1,7 @@
-import { Context } from './global-context.js';
+import { Context } from './util/global-context.js';
 import { Gate, GateType, JsonGate, Snarky } from '../snarky.js';
 import { parseHexString32 } from '../bindings/crypto/bigint-helpers.js';
-import { prettifyStacktrace } from './errors.js';
+import { prettifyStacktrace } from './util/errors.js';
 import { Fp } from '../bindings/crypto/finite-field.js';
 import { MlBool } from './ml/base.js';
