+++ src/lib/util/base58.unit-test.ts
@@ -1,7 +1,7 @@
 import { fromBase58Check, toBase58Check } from './base58.js';
-import { Test } from '../snarky.js';
+import { Test } from '../../snarky.js';
 import { expect } from 'expect';
-import { test, Random, withHardCoded } from './testing/property.js';
+import { test, Random, withHardCoded } from '../testing/property.js';
 
 let bytes = withHardCoded(
   Random.bytes(Random.nat(100)),
