+++ src/lib/util/errors.ts
