+++ src/lib/util/global-context.ts
